+++ actix/src/address/mod.rs
@@ -234,6 +234,14 @@ impl<A: Actor> fmt::Debug for WeakAddr<A> {
     }
 }
 
+impl<A: Actor> PartialEq for WeakAddr<A> {
+    fn eq(&self, other: &Self) -> bool {
+        self.wtx == other.wtx
+    }
+}
+
+impl<A: Actor> std::cmp::Eq for WeakAddr<A> {}
+
 /// The [`Recipient`] type allows to send one specific message to an actor.
 ///
 /// You can get a recipient using the `Addr::recipient()` method. It is possible
@@ -455,6 +463,7 @@ mod tests {
     }
 
     pub struct SetCounter(usize);
+
     impl Message for SetCounter {
         type Result = ();
     }