+++ actix/CHANGES.md
@@ -3,15 +3,15 @@
 ## Unreleased - 2021-xx-xx
 ### Changed
 - Updated minimum supported Rust version to 1.49.
-
-### Removed
-- Removed `Resolver` actor [#451]
 - Implement `Clone` for `WeakRecipient` [#518]
 - Extend Sender trait by a downgrade function [#518]
 - Make `WeakSender` trait public rather than crate-public [#518]
 - Add `downgrade` functionality to `Recipient` to obtain a `WeakRecipient` [#518]
 - Add `From` conversion from `Recipient` to `WeakRecipient` [#518]
+- Implement `PartialEq` and `Eq` to `WeakAddr` and `WeakAddressSender` [#523]
 
+### Removed
+- Removed `Resolver` actor [#451]
 
 ## 0.12.0 - 2021-06-06
 ### Added